sprite_viewer: HD overlays (SCV flame, Firebat spray, ...)
Two fixes needed to render iscript-spawned image_t overlays in HD:

1. Anim-num fallback in paint_hd_image. Overlay image_ts live
   outside our mapping table (which only knows unit bodies), so
   sc_r_row_for_bw_id() returns -1 and the HD path bailed out.
   Their image_id, however, corresponds 1:1 to an anim file on
   disk (anim\main_<image_id>.anim), even when images.rel marks
   the row SD-only. When the mapping-table lookup misses, fall
   back to load_sprite_layer_by_anim(image_id, "diffuse").
   Cached per image_id in a parallel cache to the body path.

2. HdFrame::offset_x/offset_y typed as int16_t, not uint16_t.
   Blizzard's anim frame table stores per-frame anchor offsets
   as signed 16-bit ints. Overlay anims (SCV flame, Firebat
   spray, siege blast, ...) routinely author negative offsets
   to anchor the atlas rect above/left of the sprite center.
   Reading them as unsigned turned -13 into 65523, which the
   paint math then multiplied through and drew the flame
   ~30000 widget pixels off-screen. Unit body frames only use
   small positive offsets, so this bug was invisible until we
   wired up the overlay path in (1).

Verified visually: SCV Walking shows engine flames; Firebat
GndAttkInit shows flame spray. The same path also applies to
other iscript-spawned overlays (spider mine tracks, casting
glows, etc.) which live in the same "orphan anim keyed by
image_id" bucket.

+++ sprite_viewer/hd_asset_loader.cpp
@@ -126,8 +126,11 @@ static bool parse_anim(const std::vector<u8>& bytes, ParsedAnim& out,
 		const u8* p = ft + i * 16;
 		f.atlas_x  = rd16le(p + 0);
 		f.atlas_y  = rd16le(p + 2);
-		f.offset_x = rd16le(p + 4);
-		f.offset_y = rd16le(p + 6);
+		// offset_{x,y} are signed int16 on disk. Reinterpret the
+		// unsigned parse via a static_cast so the bit pattern
+		// (e.g. 0xFFF3) survives as -13 rather than 65523.
+		f.offset_x = (int16_t)rd16le(p + 4);
+		f.offset_y = (int16_t)rd16le(p + 6);
 		f.w        = rd16le(p + 8);
 		f.h        = rd16le(p + 10);
 		f.flags    = rd32le(p + 12);
